(GH12) moves a lot of announcements out into their own files
