+++ announcements/rolled.js
@@ -1,0 +1,17 @@
+/* eslint-disable max-len */
+const { signedNumber } = require('../helpers/signed-number');
+
+const announceRolled = (publicChannel, channelManager, className, monster, {
+	reason,
+	roll,
+	player,
+	outcome
+}) => {
+	publicChannel({
+		announce:
+`🎲  ${player.identity} rolled a ${roll.result} (natural ${roll.naturalRoll.result}${signedNumber(roll.bonusResult)}${signedNumber(roll.modifier)}) ${reason}
+${outcome}`
+	});
+};
+
+module.exports = announceRolled;
