+++ game.js
@@ -2,14 +2,14 @@ const reduce = require('lodash.reduce');
 const moment = require('moment');
 const zlib = require('zlib');
 
-const { actionCard, monsterCard } = require('./helpers/card');
+// const { actionCard, monsterCard } = require('./helpers/card');
 const { all: cardTypes, draw } = require('./cards');
 const { all: monsterTypes } = require('./monsters');
 const { COINS_PER_VICTORY, COINS_PER_DEFEAT } = require('./helpers/coins');
 const { create: createCharacter } = require('./characters');
 const { getFlavor } = require('./helpers/flavor');
 const { globalSemaphore } = require('./helpers/semaphore');
-const { signedNumber } = require('./helpers/signed-number');
+// const { signedNumber } = require('./helpers/signed-number');
 const { XP_PER_VICTORY, XP_PER_DEFEAT } = require('./helpers/experience');
 const aws = require('./helpers/aws');
 const BaseClass = require('./baseClass');
@@ -18,8 +18,19 @@ const ChannelManager = require('./channel');
 const PlayerHandbook = require('./player-handbook');
 const Ring = require('./ring');
 
+const announceCard = require('./announcements/card.js');
+const announceCardDrop = require('./announcements/cardDrop.js');
+const announceContestant = require('./announcements/contestant.js');
+const announceDeath = require('./announcements/death.js');
+const announceEffect = require('./announcements/effect.js');
+const announceEndOfDeck = require('./announcements/endOfDeck.js');
 const announceHit = require('./announcements/hit.js');
-
+const announceLeave = require('./announcements/leave.js');
+const announceRolled = require('./announcements/rolled.js');
+const announceRolling = require('./announcements/rolling.js');
+const announceStay = require('./announcements/stay.js');
+const announceTurnBegin = require('./announcements/turnBegin.js');
+const announceXPGain = require('./announcements/xpGain.js');
 
 const PUBLIC_CHANNEL = 'PUBLIC_CHANNEL';
 
@@ -81,35 +92,42 @@ class Game extends BaseClass {
 
 	initializeEvents () {
 		const events = [
-			{ event: 'card.effect', listener: this.announceEffect },
-			{ event: 'card.miss', listener: this.announceMiss },
+			{ event: 'card.played', listener: announceCard },
+			{ event: 'cardDrop', listener: announceCardDrop },
+			{ event: 'ring.add', listener: announceContestant },
+			{ event: 'creature.die', listener: announceDeath },
+			{ event: 'card.effect', listener: announceEffect },
+			{ event: 'ring.endOfDeck', listener: announceEndOfDeck },
 			{ event: 'creature.hit', listener: announceHit },
+			{ event: 'creature.leave', listener: announceLeave },
+			{ event: 'card.rolled', listener: announceRolled },
+			{ event: 'card.rolling', listener: announceRolling },
+			{ event: 'card.stay', listener: announceStay },
+			{ event: 'ring.turnBegin', listener: announceTurnBegin },
+			{ event: 'ring.gainedXP', listener: announceXPGain },
+			{ event: 'gainedXP', listener: announceXPGain },
+
+			{ event: 'card.miss', listener: this.announceMiss },
+
 			{ event: 'card.narration', listener: this.announceNarration },
-			{ event: 'card.played', listener: this.announceCard },
-			{ event: 'card.rolled', listener: this.announceRolled },
-			{ event: 'card.rolling', listener: this.announceRolling },
-			{ event: 'card.stay', listener: this.announceStay },
-			{ event: 'cardDrop', listener: this.announceCardDrop },
-			{ event: 'creature.die', listener: this.announceDeath },
-			{ event: 'creature.leave', listener: this.announceLeave },
+
+
 			{ event: 'creature.modifier', listener: this.announceModifier },
-			{ event: 'gainedXP', listener: this.announceXPGain },
-			{ event: 'ring.add', listener: this.announceContestant },
+
+
 			{ event: 'ring.bossWillSpawn', listener: this.announceBossWillSpawn },
-			{ event: 'ring.endOfDeck', listener: this.announceEndOfDeck },
+
 			{ event: 'ring.fight', listener: this.announceFight },
 			{ event: 'ring.fightConcludes', listener: this.announceFightConcludes },
-			{ event: 'ring.gainedXP', listener: this.announceXPGain },
-			{ event: 'ring.remove', listener: this.announceContestantLeave },
-			{ event: 'ring.roundComplete', listener: this.announceNextRound },
-			{ event: 'ring.turnBegin', listener: this.announceTurnBegin },
-		]
 
-		events.map((event) => {
-			this.on(event.event, (...args) => {
-				event.listener(this.publicChannel, this.channelManager, ...args);
-			})
-		});
+			{ event: 'ring.remove', listener: this.announceContestantLeave },
+			{ event: 'ring.roundComplete', listener: this.announceNextRound }
+
+		];
+
+		events.map(event => this.on(event.event, (...args) => {
+			event.listener(this.publicChannel, this.channelManager, ...args);
+		}));
 
 		this.on('creature.heal', (...args) => {
 			this.announceHeal(this.publicChannel, this.channelManager, this.ring, ...args);
@@ -119,167 +137,6 @@ class Game extends BaseClass {
 		this.on('creature.loss', this.handleLoser);
 		this.on('creature.permaDeath', this.handlePermaDeath);
 		this.on('creature.fled', this.handleFled);
-
-	}
-
-	/* eslint-disable max-len */
-	announceXPGain (publicChannel, channelManager, className, game, {
-		contestant,
-		creature,
-		xpGained,
-		killed,
-		coinsGained
-	}) {
-		const { channel, channelName } = contestant;
-
-		let coinsMessage = '';
-		if (coinsGained) {
-			coinsMessage = ` and ${coinsGained} coins`;
-		}
-
-		let killedMessage = '';
-		if (killed) {
-			killedMessage = ` for killing ${killed.length} ${killed.length > 1 ? 'monsters' : 'monster'}.`;
-		}
-
-		channelManager.queueMessage({
-			announce: `${creature.identity} gained ${xpGained} XP${killedMessage}${coinsMessage}`,
-			channel,
-			channelName
-		});
-	}
-
-	announceCardDrop (publicChannel, channelManager, className, game, { contestant, card }) {
-		const { channel, channelName } = contestant;
-
-		const cardDropped = actionCard(card);
-
-		const announce = `${contestant.monster.identity} finds a card for ${contestant.character.identity} in the dust of the ring:
-
-${cardDropped}`;
-
-		channelManager.queueMessage({
-			announce,
-			channel,
-			channelName
-		});
-
-		channel({
-			announce
-		});
-	}
-
-	announceCard (publicChannel, channelManager, className, card, { player }) {
-		const cardPlayed = actionCard(card);
-
-		publicChannel({
-			announce:
-`${player.identity} lays down the following card:
-${cardPlayed}`
-		});
-	}
-
-	announceTurnBegin (publicChannel, channelManager, className, ring, { contestant }) {
-		const { monster } = contestant;
-
-		publicChannel({
-			announce:
-`*It's ${contestant.character.givenName}'s turn.*
-
-${contestant.character.identity} plays the following monster:
-${monsterCard(monster, contestant.lastMonsterPlayed !== monster)}`
-		});
-
-		contestant.lastMonsterPlayed = monster;
-	}
-
-	announceEndOfDeck (publicChannel, channelManager, className, ring, { contestant }) {
-		const { monster } = contestant;
-
-		publicChannel({
-			announce:
-`${monster.identity} is out of cards.`
-		});
-	}
-
-	announceNextRound (publicChannel, channelManager, className, ring, { round }) {
-		publicChannel({
-			announce:
-`
-🏁       round ${round} complete
-
-###########################################`
-		});
-	}
-
-	announceDeath (publicChannel, channelManager, className, monster, { assailant, destroyed }) {
-		let announce;
-
-		if (destroyed) {
-			announce = `${monster.identityWithHp} has been sent to the land of ${monster.pronouns[2]} ancestors by ${assailant.identityWithHp}
-
-			☠️  R.I.P ${monster.identity}
-`;
-		} else {
-			announce = `💀  ${monster.identityWithHp} is killed by ${assailant.identityWithHp}
-`;
-		}
-
-		publicChannel({ announce });
-	}
-
-	announceLeave (publicChannel, channelManager, className, monster, { assailant }) {
-		publicChannel({
-			announce:
-`${monster.identityWithHp} flees from ${assailant.identityWithHp}
-`
-		});
-	}
-
-	announceStay (publicChannel, channelManager, className, monster, { fleeRoll, player, target }) {
-		if (fleeRoll) {
-			publicChannel({
-				announce:
-	`${player.identityWithHp} tries to flee from ${target.identityWithHp}, but fails!`
-			});
-		} else {
-			publicChannel({
-				announce:
-	`${player.identityWithHp} bravely stays in the ring.`
-			});
-		}
-	}
-
-	announceRolling (publicChannel, channelManager, className, monster, {
-		reason,
-		roll,
-		player
-	}) {
-		let title = roll.primaryDice;
-		if (roll.bonusDice) {
-			title += signedNumber(roll.bonusDice);
-		}
-		if (roll.modifier) {
-			title += signedNumber(roll.modifier);
-		}
-
-		publicChannel({
-			announce:
-`🎲  ${player.identity} rolls ${title} ${reason}`
-		});
-	}
-
-	announceRolled (publicChannel, channelManager, className, monster, {
-		reason,
-		roll,
-		player,
-		outcome
-	}) {
-		publicChannel({
-			announce:
-`🎲  ${player.identity} rolled a ${roll.result} (natural ${roll.naturalRoll.result}${signedNumber(roll.bonusResult)}${signedNumber(roll.modifier)}) ${reason}
-    ${outcome}`
-		});
 	}
 
 	announceModifier (publicChannel, channelManager, className, monster, {
@@ -294,6 +151,16 @@ ${monsterCard(monster, contestant.lastMonsterPlayed !== monster)}`
 		publicChannel({
 			announce:
 `${monster.identity}'s ${attr} ${dir} by ${Math.abs(amount)}`
+		});
+	}
+
+	announceNextRound (publicChannel, channelManager, className, ring, { round }) {
+		publicChannel({
+			announce:
+`
+🏁       round ${round} complete
+
+###########################################`
 		});
 	}
 
@@ -330,28 +197,8 @@ ${monster.icon}  ${monster.givenName} now has ${monster.hp}HP.`
 		});
 	}
 
-	announceEffect (publicChannel, channelManager, className, card, {
-		player, target, effectResult
-	}) {
-		publicChannel({
-			announce:
-`${target.icon}  ${target.givenName} is currently ${effectResult} by ${player.icon}  ${player.givenName}
-`
-		});
-	}
-
 	announceNarration (publicChannel, channelManager, className, card, { narration }) {
 		publicChannel({ announce: narration });
-	}
-
-	announceContestant (publicChannel, channelManager, className, ring, { contestant }) {
-		const { character, monster } = contestant;
-
-		publicChannel({
-			announce:
-`A${getFlavor('monsterAdjective')} ${monster.creatureType} has entered the ring at the behest of ${character.icon}  ${character.givenName}.
-${monsterCard(monster)}`
-		});
 	}
 
 	announceContestantLeave (publicChannel, channelManager, className, ring, { contestant }) {
@@ -386,8 +233,6 @@ ${monsterCard(monster)}`
 
 		channelManager.sendMessages();
 	}
-	/* eslint-enable max-len */
-
 
 	handleWinner (className, monster, { contestant }) {
 		// Draw a card, maybe kick off more events (that could be messaged)
